refactor(qa): correct import paths to new broker-client
