+++ qa/integration-tests/src/test/java/io/camunda/zeebe/it/gateway/GatewayHealthIndicatorsIntegrationTest.java
@@ -12,9 +12,9 @@ import static org.assertj.core.data.Offset.offset;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import io.camunda.zeebe.broker.client.api.BrokerClusterState;
 import io.camunda.zeebe.gateway.health.Status;
 import io.camunda.zeebe.gateway.impl.SpringGatewayBridge;
-import io.camunda.zeebe.gateway.impl.broker.cluster.BrokerClusterState;
 import io.camunda.zeebe.gateway.impl.probes.health.ClusterAwarenessHealthIndicator;
 import io.camunda.zeebe.gateway.impl.probes.health.PartitionLeaderAwarenessHealthIndicator;
 import io.camunda.zeebe.gateway.impl.probes.health.StartedHealthIndicator;
